metaforge/requirements.txt

sqlalchemy
pydantic
PyYAML
